sharetasks

+++ sharetasks/public/task/edit-task.php
@@ -6,6 +6,7 @@ Auth::requireLogin();
 $method = $_SERVER["REQUEST_METHOD"];
 $errors = [];
 $data = [];
+$data["update_user"] = Auth::getUser();
 $mode = "edit";
 
 if (isset($_GET["task_id"])) {
@@ -22,6 +23,7 @@ if ($method == "GET") {
 }
 
 if (isset($_POST["save"])) { 
+    $_POST = filter_input_array(INPUT_POST, FILTER_SANITIZE_FULL_SPECIAL_CHARS);
 
     foreach($_POST as $key => $value) { 
         $data[$key] = $value;
@@ -33,18 +35,26 @@ if (isset($_POST["save"])) {
     $attaches = $attachService->getByTaskId($taskId);
     $data["attaches"] = $attaches;
 
-    // TODO: 入力チェック
     $taskService = new TaskService();
-    $taskService->update($taskId, $data);
+    $errors = $taskService->validate($data);
 
-    if (isset($data["delete_attach_ids"])) {
-        $attachService->delete($data["delete_attach_ids"]);
+    if (empty($errors)) {
+        $taskService->update($taskId, $data);
+
+        if (isset($data["delete_attach_ids"])) {
+            $attachService->delete($data["delete_attach_ids"]);
+        }
+
+        $attachService->uploadFile($taskId, $data["upload_file"]);
+
+        Util::registerMessage("task_edited", "タスクを更新しました。");
+        Util::redirect("/task/home");
+    } else {
+        $globalErrors[] = "入力内容を見直してください。";
+        if (AttachService::fileAttached($data["upload_file"])) {
+            $globalErrors[] = "添付ファイルは再選択が必要です。ご注意ください。";
+        }
     }
-
-    $attachService->uploadFile($taskId, $data["upload_file"]);
-
-    Util::registerMessage("task_edited", "タスクを更新しました。");
-    Util::redirect("/task/home");
 }
 
 if (isset($_POST["delete"])) {
@@ -62,8 +72,8 @@ if (isset($_POST["delete"])) {
 
 <div class="container">
     <?php require_once APPROOT . "/includes/shared/global-error.php"; ?>
-    <div class="card">
-        <div class="card-header">Edit Article</div>
+    <div class="card mb-3">
+        <div class="card-header">タスクの更新</div>
         <div class="card-body">
             <?php require_once APPROOT . "/includes/task/task-form.php"; ?>
         </card>
